requirements.txt
euclid3==0.1
iniconfig==2.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pluggy==1.4.0
ply==3.11
prettytable==0.7.2
pypng==0.0.19
pytest==8.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
six==1.16.0
solidpython==1.1.3
tzdata==2024.1
